Add comic and book viewer to dark theme
The changes adds the comic viewer, book viewer and the book table of contents pop up to the dark theme. It includes dark background for all, light buttons, and light links to each chapter in the table of contents. The only thing I couldn't figure out is how to get the "X" close buttons to turn blue for the comic book viewer and the book table of contents pop up.